Make the dance graph discrete and stop rebuilds from wiping it
Setup Ptaku Character deleted PtakuCharacter.controller and rebuilt a
single-clip Dance state with no DanceIndex parameter, so running that
builder silently removed the radial menu for Ptaku. It now delegates to
DanceAnimationSetup.ConfigureSharedDanceGraph after recreating the
controller, and no longer authors a Dance state or clip of its own.
Ptaku_Dance.fbx was its only consumer and is deleted; nothing else
referenced the asset by GUID or by name.

Replace the Simple1D blend tree with one state per dance and switch
DanceIndex from Float to Int. The dances are discrete choices, not a
continuum: the blend tree only produced a clean pose because SetFloat
happened to write exact integers, and adding damping later would have
cross-faded two dances together. Any State entry conditions now compare
DanceIndex with Equals, and a stale index self-corrects because the
transition to the matching state fires as soon as the index arrives.

Remove four leftover Animators on Visuals/Malpa, Wieprz, Jak and Karton.
They had no controller, applyRootMotion on and CullUpdateTransforms, and
were left over from FBX import; Ptaku never had one. No code enumerates
child Animators - the root Animator drives every character.

Set PanelSettings match to 0.5. With match 0 the UI scaled off screen
width alone, so on an ultrawide display everything inflated vertically
against an unchanged screen height.

+++ Assets/Scripts/Gameplay/PlayerAnimationDriver.cs
@@ -87,8 +87,8 @@ namespace InterrogationRoom.Gameplay
 
         public void SetDanceIndex(int selectedDance)
         {
-            if (HasParameter(DanceIndexParameter, AnimatorControllerParameterType.Float))
-                animator.SetFloat(DanceIndexParameter, selectedDance);
+            if (HasParameter(DanceIndexParameter, AnimatorControllerParameterType.Int))
+                animator.SetInteger(DanceIndexParameter, selectedDance);
         }
 
         public void PlayPunch(int variant)
